Moved folder and renamed component to EditPersonModal
